28/01/2024 0:50

+++ app/src/main/java/com/jluqgon214/sieteymedio/SieteyMedio/data/Carta.kt
@@ -3,8 +3,7 @@ package com.jluqgon214.sieteymedio.SieteyMedio.data
 class Carta(
     var nombre: Naipes,
     var palo: Palos,
-    var puntos: Int,
+    var puntos: Double,
     var idDrawable: Int
 ) {
-
 }
